// sourcecode/BTrak.Packages/snova-asset-module/Assets-module-components/assets-components/src/lib/snovasys-asset-management/models/location-management.ts
export class LocationManagement {
    seatingId: string;
    employeeId: string;
    employeeName: string;
    employeeProfileImage: string;
    seatCode: string;
    description: string;
    comment: string;
    createdDateTime: Date;
    pageSize: number;
    pageNumber: number;
    searchText: string;
    isArchived: boolean;
    sortBy: string;
    sortDirectionAsc: boolean;
    totalCount: number;
    timeStamp: any;
    branchId: string;
    entityId: string;
}